Rename component file with correct capitalization
