+++ src/components/categories/categoriesList.jsx
@@ -5,8 +5,8 @@
 import { useNavigate } from "react-router-dom"
 import { useEffect, useState } from "react"
 import { getAllCategories, deleteCategory } from "../../managers/CategoryManager";
-import { DeleteButton } from "../buttons/deleteButton";
-import { EditButton } from "../buttons/editButton";
+import { DeleteButton } from "../buttons/DeleteButton";
+import { EditButton } from "../buttons/EditButton";
 
 const initialCategoryState = {
     id: 0,
