+++ src/components/posts/MyPosts.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react"
 import { getPostsByUserId, deletePost } from "../../managers/PostManager"
 import { useNavigate } from "react-router-dom"
-import { EditButton } from "../buttons/editButton"
+import { EditButton } from "../buttons/EditButton"
 // import "./MyPosts.css"
 
 export const MyPosts = ({ token }) => {
