+++ src/components/posts/Posts.jsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { getPostById } from "../../managers/PostManager";
 import { CommentForm } from "../comments/comments";
-import { EditButton } from "../buttons/editButton";
+import { EditButton } from "../buttons/EditButton";
 // import "./Post.css"
 
 export const PostDetails = () => {
